webnn/clamp2.html

<!doctype html>
<meta charset=utf-8>
<title>test clamp operation</title>
<link rel="help" href="https://webmachinelearning.github.io/webnn/#api-mlgraphbuilder-clamp">
<script src=../resources/testharness.js></script>
<script src=../resources/testharnessreport.js></script>
<script src="./webnn-polyfill.js"></script>
<div id=log></div>
<script type="module">
  'use strict';
  import * as utils from './utils.js';
  let builder;
  setup(() => {
    const context = navigator.ml.createContext();
    builder = new MLGraphBuilder(context);
  })

  function testClamp(inputShape, inputValue, expected, options = {}) {
    const x = builder.input('x', {type: 'float32', dimensions: inputShape});
    const y = builder.clamp(x, options);
    const graph = builder.build({y});
    const inputs = {'x': new Float32Array(inputValue)};
    const outputs = {'y': new Float32Array(utils.sizeOfShape(inputShape))};
    graph.compute(inputs, outputs);
    assert_array_approx_equals_ulp('clmap', outputs.y, expected.float32, 0);
  }

const inputData = [
  -3.4356449350874696, -6.530945988411405,  -8.175760663838268, 2.0879641317522726,
  -4.480150236948526,  -8.591504561715722,  5.071455429211573,  -6.618697702258771,
  4.224577823136105,   6.450272349350044,   -8.799923845835664, -3.3445965406946643,
  5.550524270215341,   1.2788677438688012,  9.333702625514768,  9.2261637863086,
  -7.302720212371034,  1.7865902395032585,  5.564981581526375,  3.145101011211482,
  -8.275078596251655,  -1.3557080837143296, 7.348269585030259,  -5.530012756488021,
];
// expected data by clamping input data with default options
const expectedDefault = {
  float32: [
    -3.4356448650360107, -6.530945777893066, -8.175760269165039, 2.0879640579223633,
    -4.48015022277832,   -8.591504096984863, 5.071455478668213,  -6.618697643280029,
    4.224577903747559,   6.450272560119629,  -8.79992389678955,  -3.3445966243743896,
    5.5505242347717285,  1.2788677215576172, 9.33370304107666,   9.226163864135742,
    -7.302720069885254,  1.7865902185440063, 5.564981460571289,  3.1451010704040527,
    -8.275078773498535,  -1.355708122253418, 7.348269462585449,  -5.530012607574463,
  ],
};
// expected data by clamping input data within [-5, 7]
const expectedMinMax = {
  float32: [
    -3.4356448650360107, -5,                 -5,                2.0879640579223633,
    -4.48015022277832,   -5,                 5.071455478668213, -5,
    4.224577903747559,   6.450272560119629,  -5,                -3.3445966243743896,
    5.5505242347717285,  1.2788677215576172, 7,                 7,
    -5,                  1.7865902185440063, 5.564981460571289, 3.1451010704040527,
    -5,                  -1.355708122253418, 7,                 -5,
  ],
};
// expected data by clamping input data within [0, 6]
const expectedRelu6 = {
  float32: [
    0,                  0,                  0,                 2.0879640579223633,
    0,                  0,                  5.071455478668213, 0,
    4.224577903747559,  6,                  0,                 0,
    5.5505242347717285, 1.2788677215576172, 6,                 6,
    0,                  1.7865902185440063, 5.564981460571289, 3.1451010704040527,
    0,                  0,                  6,                 0,
  ],
};
// expected data by clamping input data with specified minValu=-5
const expectedMin = {
  float32: [
    -3.4356448650360107, -5,                 -5,                2.0879640579223633,
    -4.48015022277832,   -5,                 5.071455478668213, -5,
    4.224577903747559,   6.450272560119629,  -5,                -3.3445966243743896,
    5.5505242347717285,  1.2788677215576172, 9.33370304107666,  9.226163864135742,
    -5,                  1.7865902185440063, 5.564981460571289, 3.1451010704040527,
    -5,                  -1.355708122253418, 7.348269462585449, -5,
  ],
};
// expected data by clamping input data with specified minValu=0
const expectedMinZero = {
  float32: [
    0,                  0,                  0,                 2.0879640579223633,
    0,                  0,                  5.071455478668213, 0,
    4.224577903747559,  6.450272560119629,  0,                 0,
    5.5505242347717285, 1.2788677215576172, 9.33370304107666,  9.226163864135742,
    0,                  1.7865902185440063, 5.564981460571289, 3.1451010704040527,
    0,                  0,                  7.348269462585449, 0,
  ],
};
// expected data by clamping input data with specified maxValu=7
const expectedMax = {
  float32: [
    -3.4356448650360107, -6.530945777893066, -8.175760269165039, 2.0879640579223633,
    -4.48015022277832,   -8.591504096984863, 5.071455478668213,  -6.618697643280029,
    4.224577903747559,   6.450272560119629,  -8.79992389678955,  -3.3445966243743896,
    5.5505242347717285,  1.2788677215576172, 7,                  7,
    -7.302720069885254,  1.7865902185440063, 5.564981460571289,  3.1451010704040527,
    -8.275078773498535,  -1.355708122253418, 7,                  -5.530012607574463,
  ],
};
// expected data by clamping input data with specified maxValu=0
const expectedMaxZero = {
  float32: [
    -3.4356448650360107, -6.530945777893066, -8.175760269165039, 0,
    -4.48015022277832,   -8.591504096984863, 0,                  -6.618697643280029,
    0,                   0,                  -8.79992389678955,  -3.3445966243743896,
    0,                   0,                  0,                  0,
    -7.302720069885254,  0,                  0,                  0,
    -8.275078773498535,  -1.355708122253418, 0,                  -5.530012607574463,
  ],
};

// test = {
//   purpose: [
//     [inputShape, expectedData, options],
//   ],
// };
const tests = {
  'default options': [
    // 1D
    [[24], expectedDefault],
    // 2D
    [[4, 6], expectedDefault],
    // 3D
    [[2, 3, 4], expectedDefault],
    // 4D
    [[2, 3, 2, 2], expectedDefault],
    // 5D
    [[2, 3, 2, 1, 2], expectedDefault],
  ],
  'only specified minValue option': [
    // 1D
    [[24], expectedMin, {minValue: -5}],
    [[24], expectedMinZero, {minValue: 0}],
    // 2D
    [[4, 6], expectedMin, {minValue: -5}],
    [[4, 6], expectedMinZero, {minValue: 0}],
    // 3D
    [[2, 3, 4], expectedMin, {minValue: -5}],
    [[2, 3, 4], expectedMinZero, {minValue: 0}],
    // 4D
    [[2, 3, 2, 2], expectedMin, {minValue: -5}],
    [[2, 3, 2, 2], expectedMinZero, {minValue: 0}],
    // 5D
    [[2, 3, 2, 1, 2], expectedMin, {minValue: -5}],
    [[2, 3, 2, 1, 2], expectedMinZero, {minValue: 0}],
  ],
  'only specified maxValue option': [
    // 1D
    [[24], expectedMax, {maxValue: 7}],
    [[24], expectedMaxZero, {maxValue: 0}],
    // 2D
    [[4, 6], expectedMax, {maxValue: 7}],
    [[4, 6], expectedMaxZero, {maxValue: 0}],
    // 3D
    [[2, 3, 4], expectedMax, {maxValue: 7}],
    [[2, 3, 4], expectedMaxZero, {maxValue: 0}],
    // 4D
    [[2, 3, 2, 2], expectedMax, {maxValue: 7}],
    [[2, 3, 2, 2], expectedMaxZero, {maxValue: 0}],
    // 5D
    [[2, 3, 2, 1, 2], expectedMax, {maxValue: 7}],
    [[2, 3, 2, 1, 2], expectedMaxZero, {maxValue: 0}],
  ],
  'both specified minValue and maxValue options': [
    // 1D
    [[24], expectedMinMax, {minValue: -5, maxValue: 7}],
    [[24], expectedRelu6, {minValue: 0, maxValue: 6}],
    // 2D
    [[4, 6], expectedMinMax, {minValue: -5, maxValue: 7}],
    [[4, 6], expectedRelu6, {minValue: 0, maxValue: 6}],
    // 3D
    [[2, 3, 4], expectedMinMax, {minValue: -5, maxValue: 7}],
    [[2, 3, 4], expectedRelu6, {minValue: 0, maxValue: 6}],
    // 4D
    [[2, 3, 2, 2], expectedMinMax, {minValue: -5, maxValue: 7}],
    [[2, 3, 2, 2], expectedRelu6, {minValue: 0, maxValue: 6}],
    // 5D
    [[2, 3, 2, 1, 2], expectedMinMax, {minValue: -5, maxValue: 7}],
    [[2, 3, 2, 1, 2], expectedRelu6, {minValue: 0, maxValue: 6}],
  ],
};

  test(() => {
    for (let purpose in tests) {
      const subTests = tests[purpose];
      for (let i = 0; i < subTests.length; i++) {
        testClamp(subTests[i][0], inputData, subTests[i][1], subTests[i][2]);
      }
    }
  }, 'clamp');
</script>
